+++ frontend/src/App.jsx
@@ -7,6 +7,7 @@ import Support from './pages/Support.jsx'
 import Account from './pages/Account.jsx'
 import Rapporten from './pages/Rapporten.jsx'
 import Meldingen from './pages/Meldingen.jsx'
+import Garantie from './pages/Garantie.jsx'
 import Admin from './pages/Admin.jsx'
 import Webshop from './pages/Webshop.jsx'
 import Apps from './pages/Apps.jsx'
@@ -398,6 +399,7 @@ export default function App() {
         <Route path="/rapporten" element={<Rapporten />} />
         <Route path="/meldingen" element={<Meldingen />} />
         <Route path="/support" element={<Support user={user} />} />
+        <Route path="/garantie" element={<Garantie />} />
         <Route path="/account" element={<Account user={user} onLogout={onLogout} />} />
         <Route path="/admin"        element={<Admin />} />
         <Route path="/onderhoud"    element={<Onderhoud />} />